Rewrite code to use simple free variable names.
The idea is to make it easy to do a textual search+replace to bind a free
variable within a code snippet. Source->source rewrites are handled by
the RenameFreeVar class; I couldn't get the clang rewriting infrastructure
to do the right thing, so RenameFreeVar maintains its own set of edits
to apply to a character buffer, and a single getRewrittenString method
that creates the fresh std::string that results from applying the edits
to the character buffer.