Add -get-scope=<unsigned int> option.
To insert a code snipped with free variables, we need to know what
variables are bound at a given insertion point so that we can do a
substitution of free -> bound variables.  This adds a clang-mutate
pass that gets the variables in scope at a given statement. Variables
are then printed out one-per-line, from the variable with the closest
definition outwards. The user provides a bound on the number of
variables to emit.