Bugfix to find directories on the source path.
Previously, we attempted to construct the full path to a directory
in the dwarf-dump output using realpath.  However, in this case,
only the current working directory is utilized for the source path.
We also need to account for when the directory can be found relative
to the compilation unit directory.  ok self@ IP:VET-TA2