Handle statements that span several buffers.
There is an obviously inefficient place where I figured out the
character delta between two source locations by getting a string
representing the source in that interval and taking its length.
I'm sure there is a more efficient way to get the length of a
CharSourceRange, but I couldn't find it.