Don't rename DeclRef ASTs underneath macro tokens.
The rewriting buffer was getting all garbled because of attempting to
apply rewrites to DeclRefs that are hidden inside of a macro expansion.
clang will tell you that the source location for the declref is valid,
but it turns out to have the wrong position and size -1.  Skip these.
This may result in extra free variables appearing that are not reflected
in the snippet, but that shouldn't affect mutations anyway.