Make -swap fail on overlapping statements.
If the source ranges for the arguments to -swap overlap, there is
not a very good choice for what we should do; just trying to forge
on anyhow causes the rewriter to segfault.  This change checks if
the ranges overlap before doing the rewriting and, if so, prints out
and error message and fails.