Extract macros from snippets and upgrade clang version.
This change threads around a clang CompilerInstance so that we can get
access to the preprocessor and other goodies during listing generation
or mutation. It uses this to figure out what macros are referenced in
an expression and recover the relevant macro definitions. There seems to
be a bug with nested macros, where the inner macro isn't detected.

Also update clang-mutate to use the latest LLVM version; this mainly
involved changing the statement pretty-printing, slightly tweaking
the top-level tool creation, and modifying the libraries to link
against.