Add type entries to the JSON database.
This adds a representation of types to the JSON database.  typedefs,
structs, and unions are all currently added, along with any such types
that they transitively depend on. If a type is defined in a system
header, the corresponding database entry will just say "hey, include
this file" rather than giving the type's definition.  The list of
paths to use as possible prefixes for system headers is currently just
baked in.