Remember children of blocks in json output.
To support manipulations that may require gathering several statements
in a row from a block, have each CompoundStmt entry in the json output
remember the counters for its children.  Combined with the parent_counter
field, this lets us walk up from any AST to its enclosing full statement
and then to the next statement in that block.