Report the index along with each free variable.
Define the "index" of a free variable in a statement to be the number of
scopes between the statement and the variable's declaration.  Examples:

{
  int x;
  x = 17;   // x has index 0 in this statement
  {
    int y;
    x = y; // x has index 1 in this statement, y has index 0
  }
}

The index of a free variable is needed when deciding how to uniformly
rename free variables on a set of statements.  It may also be useful
information for rebinding names during mutation; for example, when
inserting a snippet we may want to bind free variables to variables
with approximately the same index (think of a for-loop iterating
over an array passed as a function parameter)