Show declaration index in -get-scope.
When rebinding variables for chunks of statements that cross
the end of a scope, we need to keep track of the index of each
free variable so that a variable is only bound to declarations
that are in scope at each of its uses.  This changes the -get-scope
output to report the declared variables in each scope, from nearest
outward.