Backing out ; in json snippets, adding full_stmt.
Removes the change that added trailing semicolons in the json snippets.

Adds a full_stmt field to the json snippets that is true when the
parent of a statement is of class CompoundStmt.  This is correlated
with Jason's is_complete_c_stmt field but *not* equivalent; for example,
in "while (X) { Y; }" both X and Y are complete statements, but only
Y is a "full statement".