Add function prototypes to the JSON database.
This change adds JSON entries corresponding to function
prototypes.  There is one such entry for each FunctionDecl
with a body that is traversed by the ASTLister.  The prototype
entries include information about parameter and return types
(referencing the type database by hash), the statement corresponding
to the function body, and the text of the prototype itself.