Fix a bug with forward-declared functions.
It seems that (almost?) every function declaration in clang will have
a body, even for forward declarations. In those cases, the body will
the be body of the actual definition, so the idea of grabbing all text
from the start of the declaration to the end of the definition pulls
in everything between the forward declaration and the eventual definition.
This fixes it so that RegisterFunctionDecl is a no-op except for when we
are processing the actual definition of a function.