Join address ranges over a line range.
For some statements (e.g. blocks, compound statements), the last line
of the statement probably does not have an entry in the DWARF line
table; this table only shows the start of statements, from what I can
tell. Previously, this would mean that we fail to assign bytes to
many compound statements.

This changes the behavior so that to find the bytes associated to
a statement, we gather all of the address intervals corresponding to
lines in the line range of the statement. Then, the address intervals
are joined to get the smallest interval that covers them all; this is
returned as the address interval for the statement.

We will now be over-approximating the set of bytes associated to some
statements (rarely), but I think this is better than failing to
associate bytes to certain statements (commonly).