Ensure all statements have a unique counter.
The spine is a map of Stmt* -> Counter.  When memory for a new statement
is allocated to the same location as a previous statement, statements
in the spine are overwritten, and thus, there may be fewer statements
in the spine than we have actually visited.  The ASTEntryList contains
all of the statements which have been listed, so use this for counting
purposes.