Fix for 2-point crossover bug.
-set-range was incorrectly sorting stmt1 and stmt2 to put the lower AST
counter first; there is no reason for this, and it causes clang.lisp's
crossover methods to generate junk.

Fixing the crossover also means that the semicolon normalization of
text needs to be a little bit smarter; now, instead of only checking
the last character it will at least scan backwards from the end to
the first non-whitespace character.