Fix bad src_text due to free vars in macros.
If a macro's expansion text includes a reference to a variable
that is not bound in a snippet, we were still trying to generate
replacement text. But the offsets were coming out 0, so the
replacement would end up at the start of the overall snippet.

We don't really have a good solution for creating variants of
macros in order to make the expanded reference modifiable, so
this will still generate text snippets that may make further
mutation difficult. But at least it won't be garbage.