Correctly add/remove commas when mutating lists.
This change refactors and generalizes the ;- and {}-adding code.
ASTs now have a syn_ctx field, of type SyntacticContext, that knows
about the proper delimiter syntax before, at, and after the AST's text.
This information is plumbed to the rewriting operations, which then
use it to adjust the mutation text as needed.

The syntactic context of an AST is added to its Requirements, so new
contexts can be created by adjusting the AST and adding a new
implementation to SyntacticContext.cxx.