Use SyntacticContext::Braced for CompoundStmts.
This ensures that when you replace a braced if body with a single
unbraced statement, the braces are kept.

The other alternative, which may be better in the long term, is to build
in some understanding of the structure of if statements (and do/while
loops). But that is a bigger project to implement.