Fix source range for ASTs within a macro expansion
For any AST that is generated by a macro, the source range will
correspond to the text of the macro invocation. Previously, that range
would sometimes extend to the end of the parent statement.

One small problem remains: if the macro invocation is followed by a
semicolon, the ASTs within the macro will only include the semicolon in
their normalized range if they are full statements. This can lead to
semicolons in some cases.