Add field for CFG successors.
This is somewhat expensive so it's only generated if the -cfg flag is
passed.

Mapping clang's CFG onto statements is a little tricky. The last
statement of each block is succeeded by the first statement of each
successor block. All other statements are succeeded by the last
statement of their block. This is good enough for my current purposes.