Prevent crash from invalid type location.
Built-in struct types such as va_list may not have a valid source
location. The filename of an invalid PresumedLoc is a NULL pointer,
which causes a crash when we try to convert it to std::string.

Prevent this by checking for invalid locations and using an empty string
for the filename.