Adjust RenameFreeVars offsets to handle macros.
RenameFreeVars generates replacements which store offsets from the start
of the AST text. But the Ast constructor can adjust its source range to
account for macros, which breaks the replacements. So we have to generte
the replacements after constructing the Ast, using its final source
range.

I also removed code that was generating the replacement text and storing
it in the Requirements before creating the Ast. This text is no longer
used.