Replace strcmp-macro test.
This test was indended to ensure all ASTs within a macro
expansion had the same end offset, even if the macro declares
a variable.  As strcmp is not defined the same way on all
platforms, this test would fail on some machines.  To remedy this,
we use a non-system, user-defined macro as test data.