Fix end offset for nested macro expansions.
Previously, the end offset for macros containing a
nested inner macro in some instances would not be correct.
The end offset would point to the end of the macro name,
not the end of the macro expansion.  This change protects
against this and helps ensure the `begin_off` field is always
less than or equal to the `end_off` field.